Add note about why visual tests are slow in Jest
The goal was to run the acceptance and visual tests in parallel with
Jest and a pool of browsers connected to the hub using
`docker-compose scale`. Nonetheless the code is in a much better place
now so I'm super happy.

Even without visual tests Jest is still slower than Mocha (4s vs 3s).